feat: add SeqOverlapPatternFields component for managing sequence overlap patterns
