+++ docs/static/swagger-versions/index.json
@@ -1,4 +1,6 @@
 [
+  "v1.50.0"
+,
   "v1.49.0"
 ,
   "v1.48.1"
