+++ docs/static/swagger-versions/v1.50.0.json
